performance_agent/customerConversationPerformanceAnalysis/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Configuration dependencies
PyYAML>=6.0

# LLM and Agent dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# SAP AI Core dependencies
requests>=2.31.0
requests-oauthlib>=1.3.0
PyJWT>=2.0.0

# MongoDB dependencies
pymongo>=4.6.0

# Development dependencies
pytest>=6.0
pytest-cov>=2.0
black>=21.0
flake8>=3.9
mypy>=0.910
isort>=5.9
httpx>=0.24.0  # for testing FastAPI endpoints
